hclsyntax: ConditionalExpr unification fallback
If we are unable to unify the two sides of a conditional expression,
this should not result in an error. Instead we should return the
unconverted value, and if this causes an error higher up in the
evaluation, it can be handled there.

The motivation for this is to allow disjoint conditional expression
values to be evaluated, where the context of the expression permits
this. For example, the `merge` function can combine any object types,
and we should be able to use conditional expressions as arguments to
this function.