hclwrite: Add block label quoting to format
Blocks consist of an identifier, followed by zero or more labels, and a
body. Each label can be either an identifier or a quoted string literal.

This commit updates hclwrite's format logic to replace any identifier
labels with quoted strings. That is, for this input:

  resource null_resource test { }

hclwrite will prduce the following output:

  resource "null_resource" "test" { }