Honor hcl1 decode tags
This adds support for using hcl1 tag for decoding purposes. It still
falls back to `hcl` tag.

If an app wants to support parsing both hcl1 and hcl2 syntax, it is
beneficial to be able to set different tags with different options.  For
example, `unusedKeys` is a valid tag option in hcl1 but is not hcl2
option.