Fix the logic for `null[*]`
According to https://discuss.hashicorp.com/t/pattern-to-handle-optional-dynamic-blocks/2384/2
`null[*]` is defined and should return an empty tuple. The logic in
SplatExpr.Value() is slightly wrong thought as it returns early with a
dynamic value even when the value is actually known to be null.

The fix is easy, just check for null sources first, then for dynamic
values. I added a test for this in TestExpressionParseAndValue().