Add support for squash in hcl tags
In hcl v1, it was possible to specify ‘squash’ on embedded structures as it is with json (done by default) or yaml (with ‘inline’) or mitchellh/mapstructure (with ‘squash’). That functionality disappeared with v2.

The alternatives we found relied on the ‘remain’ tag which is not really intended to handle that situation. Moreover, there can be only one ‘remain’ tag on a structure.

The implementation support that ‘label’ could be specified anywhere in the hierarchy and the field order is maintained.

With the current implementation, it is possible to specify the same name tag to many fields. The result is that the latest override the previous one silently. At the end, the desired field is not set. We changed the code to support having multiple fields with the same tag. The new result is that all fields will get populated with the desired associated element.

Instead of handling just an index position within the structure, we changed the index to be a field path. That allows us to retrieve the right information when the struct contains squashed embedded structs.

An optional field that has a zero value should not be rendered on Encode or Marshal.

Finally, we added standard marshaling functions ‘Marshal’ and ‘Unmarshal’ into hclsimple package. This allows interaction with the library in the same way we do with json or yaml.