Fix parseQuotedStringLiteral zero end range
If a block label had an invalid quoted string literal, the parser would
previously create label ranges with uninitialized `End` values. For
Terraform, this would eventually percolate into invalid diagnostic
output due to the assumption that the end of a range is always greater
than or equal to the start.

This commit addresses this by using the terminating token's position as
the end of the range, even if it's not a closing quote.