hclsyntax: Fix incorrect examples of object expressions in the spec
In the text talking about how to resolve the ambiguity between "for"
expressions and values/keys called "for" there was previously an incorrect
statement about the behavior of "(for)" as an object key.

That situation causes HCL to take the value of a variable named "for".
To get the literal string "for" we must instead write the name in quotes,
similarly to how we'd represent a key string that includes characters
which HCL doesn't admit into an identifier.