build: Per-commit checks with GitHub Actions
We previously used CircleCI for this but we no longer have CircleCI
enabled for this repository and so we've been left with no automatic
checks.

This is an alternative check workflow implemented with GitHub Actions.
This isn't as complex/comprehensive as the CircleCI workflow that came
before it, but my initial goal here is only to have some basic automatic
feedback on new PRs as a starting point; we can expand this later if we
want to reintroduce other features such as code coverage reports.