deps: update crypto dependency for macOS support
Not sure if this is M1 specific or just newer versions of macOS altogether,
but hclfmt cannot be installed with the transitive dependency on
`golang.org/x/sys@v0.0.0-20190502175342-a43fa875dd82`
caused by the dependency on `golang.org/x/crypto`.

Update the crypto library so that a newer version of sys gets pulled in,
which compiles fine.