ext/typeexpr: [PROTOTYPE] "inferred" keyword
This is an alternative to "any" that hopefully better communicates that it
is not a dynamic typing mechanism but instead a request for automatic
type _inference_.

This should not be merged in its current form. It's here only to explore
the possibility.

In particular, a real version of this would probably not emit a
deprecation warning -- at least at first -- to avoid being immediately
disruptive to existing valid configurations. But the warning is here in
the prototype mainly to help explore the implications of generating a
warning in this context, where we previously generated only errors or no
diagnostics at all.