hclsyntax: ConditionalExpr can refine its unknown results
When ConditionalExpr has an unknown predicate it can still often infer
some refinement to the range of its result by noticing characteristics
that the two results have in common.

In all cases we can test if either result could be null and return a
definitely-not-null unknown value if not.

For two known numbers we can constrain the range to be between those two
numbers. This is primarily aimed at the common case where the two possible
results are zero and one, which significantly constrains the range.

For two known collections of the same kind we can constrain the length
to be between the two collection lengths.

In these last two cases we can also sometimes collapse the unknown into
a known value if the range gets reduced enough. For example, if choosing
between two collections of the same length we might return a known
collection of that length containing unknown elements, rather than an
unknown collection.