decoder: add hasKey tag
The `hasKey` fixes a problem where a struct with a list field that has
elements with the `key` tag receives the wrong key value when the block
is defined without a label.

For example, a struct like this:

```
type A struct {
  Bs []*B `hcl:"b"`
}

type B struct {
  C string `hcl:",key"`
  D string
}
```

And an input like this:

```
b {
  d = "e"
}

b "z" {
  d = "y"
}
```

Generates an incorrect output:

```
&A {
  Bs: []*B{
    {
      C: "d", // wut?
      D: "e",
    },
    {
      C: "z", // ok
      D: "y",
    },
  },
}
```

This happens because when the decoder is traversing the AST it doesn't
know if a nested field expects a key value, and so it makes a wrong
assumption about what the key value should be.

Internally, each item in the AST has keys and point to another node of
the tree. For blocks, the keys are the block type plus its labels. In
the example above, the first block has one key, the identifier "b". The
second block has two keys: the identifier "b" and a string label `"z"`.

When the file is processed, the decoder receives the root node and then
iterates over each field of A.

Since the first field is a list of Bs, it filters from the root the
blocks that have a "b" identifier as the first key. This filtering
process also removes the "b" entry from the item's keys.

This leaves blocks without a label (such as the first "b" above) with no
keys left. The decoder then considers it to be an exact match and moves
to the block content as the next element to process.

Blocks that do have labels (like the second "b" block) still have the
label as part of the key, and so are considered to be a prefix match.

Exact matches have their children processed, not the item itself. This
creates the problem described here because the children will have an
unrelated value as key. In the example above, the block without label
receives "d" as key because that's the first key of the sub-element.

The `hasKey` tag signals to the decoder that the list elements expect to
receive a key, even an empty one. And so this commit turns exact matches
into prefix matches, forcing the item itself to be processed first, and
add an empty key component as the label, instead of proceeding to its
children.

Adding this tag to the example generates the expected output:

```
type A struct {
  Bs []*B `hcl:"b,hasKey"`
}

type B struct {
  C string `hcl:",key"`
  D string
}
```

```
&A {
  Bs: []*B{
    {
      C: "", // fixed
      D: "e",
    },
    {
      C: "z", // ok
      D: "y",
    },
  },
}
```